Add main entry point for package.json.
Without this change, Ember CLI 0.2.0 will have the following warning for
all users:

```
% ember s

The package `ember-cli-data-factory-guy` is not a properly formatted
package, we have used a fallback lookup to resolve it at
`/some-path/goes/here/node_modules/ember-cli-data-factory-guy`. This is
generally caused by an addon not having a `main` entry point (or
`index.js`).
```